api-gateway/requirements.txt
flask==3.0.3
PyJWT==2.8.0
requests==2.32.3
werkzeug==3.0.3
